src/treatment_extraction/llm_extractor.py
"""LLM-based treatment extraction interface."""

from typing import List, Dict, Optional


class LLMExtractor:
    """LLM-based extractor for Ayurvedic treatments.
    
    Uses large language models to extract treatment information
    from Sanskrit texts with higher accuracy than rule-based methods.
    
    Note:
        Requires API access to LLM service (OpenAI, Anthropic, etc.)
        or local model deployment.
    
    Example:
        >>> extractor = LLMExtractor(model="gpt-4")
        >>> treatments = extractor.extract("ज्वरस्य निदानं तिक्ताम्ललवणं परिहरेत्")
    """
    
    def __init__(
        self,
        model: str = "gpt-3.5-turbo",
        api_key: Optional[str] = None,
        temperature: float = 0.3,
    ):
        """Initialize the LLM extractor.
        
        Args:
            model: Model identifier to use.
            api_key: API key for the LLM service.
            temperature: Sampling temperature for generation.
        """
        self.model = model
        self.api_key = api_key
        self.temperature = temperature
        self.client = None
    
    def _initialize_client(self) -> None:
        """Initialize the LLM client."""
        try:
            from langchain_openai import ChatOpenAI
            self.client = ChatOpenAI(
                model=self.model,
                api_key=self.api_key,
                temperature=self.temperature,
            )
        except ImportError:
            print("Warning: langchain-openai not installed")
            self.client = None
    
    def extract(self, text: str) -> List[Dict[str, any]]:
        """Extract treatments using LLM.
        
        Args:
            text: Input Sanskrit text.
            
        Returns:
            List of extracted treatments.
        """
        if self.client is None:
            self._initialize_client()
        
        if self.client is None:
            return self._fallback_extract(text)
        
        prompt = f"""Extract Ayurvedic treatments from the following Sanskrit text.
Return the treatments in JSON format with fields: treatment_name, description, dosage.

Text: {text}

Return:"""
        
        try:
            response = self.client.invoke(prompt)
            return self._parse_response(response.content)
        except Exception as e:
            print(f"Error extracting treatments: {e}")
            return self._fallback_extract(text)
    
    def _parse_response(self, response: str) -> List[Dict[str, any]]:
        """Parse LLM response into structured format.
        
        Args:
            response: Raw LLM response.
            
        Returns:
            List of treatments.
        """
        return [{"raw_response": response, "method": "llm"}]
    
    def _fallback_extract(self, text: str) -> List[Dict[str, any]]:
        """Fallback extraction when LLM is unavailable.
        
        Args:
            text: Input text.
            
        Returns:
            Empty list (placeholder).
        """
        return [{"error": "LLM unavailable", "text": text}]
    
    def extract_batch(self, texts: List[str]) -> List[List[Dict[str, any]]]:
        """Extract treatments from multiple texts.
        
        Args:
            texts: List of input texts.
            
        Returns:
            List of treatment lists.
        """
        return [self.extract(text) for text in texts]
    
    def extract_with_sources(
        self,
        text: str,
        source_name: str,
    ) -> Dict[str, any]:
        """Extract treatments with source tracking.
        
        Args:
            text: Input Sanskrit text.
            source_name: Name of the source text.
            
        Returns:
            Treatments with source metadata.
        """
        treatments = self.extract(text)
        return {
            "source": source_name,
            "treatments": treatments,
            "text": text,
        }


def demo():
    """Demonstration function for LLM extraction."""
    print("LLM Extractor - Placeholder")
    print("Requires API key for full functionality")
    
    extractor = LLMExtractor()
    result = extractor.extract("ज्वरस्य निदानं तिक्ताम्ललवणं परिहरेत्")
    print(result)


if __name__ == "__main__":
    demo()
