requirements.txt
# Ayurveda Sanskrit LLM - Python Dependencies

# Core
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# NLP & Transformers
transformers>=4.35.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Search
faiss-cpu>=1.7.4

# LLM & LangChain
langchain>=0.1.0
langchain-community>=0.0.10

# Utilities
tqdm>=4.66.0
PyYAML>=6.0
jsonlines>=4.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
